+++ src/main/kotlin/com/groom/auth/infrastructure/authentication/AuthenticationCoreRepository.kt
@@ -1,4 +1,4 @@
-package com.groom.auth.infrastructure
+package com.groom.auth.infrastructure.authentication
 
 import com.groom.auth.domain.authentication.Authentication
 import com.groom.auth.domain.authentication.AuthenticationRepository
